requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-core==2.1.0
google-cloud-speech==2.10.0
google-cloud-storage==1.42.3
google-cloud-videointelligence==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
srt==3.5.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
